refactor(reports): split the report aggregate into modules
aggregate.ts had grown to 809 lines and held five separate concerns. It
now keeps only the entry point, and each concern moves into the sibling
aggregate-*.ts file that already names it:

- aggregate-types.ts: the aggregate shape, shared by every module
- aggregate-calendar.ts: local dates and bucket boundaries
- aggregate-rows.ts: folding rows into the period totals
- aggregate-history.ts: the daily, weekly and monthly split
- aggregate-wallets.ts: grouping by managed wallet and role

aggregate.ts re-exports the same public surface, so no caller changes.
aggregate-amounts.ts and aggregate-serialization.ts now read the types
from aggregate-types.ts instead of from aggregate.ts, which removes the
import cycle those two had with it.

No behavior change: every moved line is identical to its original.
